requirements.txt
Django
djangorestframework
flake8
requests